+++ src/main/java/webserver/httphandler/HttpHandler.java
@@ -1,7 +1,9 @@
-package webserver;
+package webserver.httphandler;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import webserver.HttpRequest;
+import webserver.HttpResponse;
 
 import java.io.IOException;
